+++ src/adminAuth.ts
@@ -8,7 +8,8 @@
  */
 
 import { getRandomBytes, toHex } from './_internal/crypto.js'
-import { type Logger, resolveLogger } from './logger.js'
+import { resolveLogger } from './_internal/resolveLogger.js'
+import type { Logger } from './logger.js'
 import {
 	createPrincipalAuth,
 	type AuthPrincipal,